Step 1: Extract BuilderOptions and friends to types/options.ts
Move BuilderOptions, BuilderResult, BuildProgressListener,
BuildProgressSnapshot, and BuildProgressStartPayload from
builder/builder.ts into types/options.ts (Layer 0 leaf).
builder/builder.ts re-exports them for backward compat.

This is prerequisite for breaking the plugins/types.ts <->
builder/builder.ts cycle in subsequent steps.
